+++ docs/tutorial/axes_title_tooltip.js
@@ -62,6 +62,15 @@ Promise.all([dataLoaded, mdChartLoaded]).then((results) => {
     },
     {
       anims: [
+        (chart) =>
+          chart.animate({
+            subtitle: "with fancy animations",
+            caption: "Source: Vizzu tutorial",
+          }),
+      ],
+    },
+    {
+      anims: [
         (chart) => {
           chart.feature('tooltip', true)
           return chart