+++ src/apps/weblib/js-api/events.js
@@ -11,18 +11,20 @@ export default class Events {
     }
 
     if (!this.eventHandlers.has(eventName)) {
-      const func = (param) => {
-        this._invoke(eventName, param)
+      let cfunc = null
+      if (!this._isJSEvent(eventName)) {
+        const func = (eventPtr, param) => {
+          this._invoke(eventName, param, eventPtr)
+        }
+        cfunc = this.module.addFunction(func, 'vii')
+        const cname = this.vizzu._toCString(eventName)
+        try {
+          this.vizzu._callOnChart(this.module._addEventListener)(cname, cfunc)
+        } finally {
+          this.module._free(cname)
+        }
       }
-      const cfunc = this.module.addFunction(func, 'vi')
-      const cname = this.vizzu._toCString(eventName)
       this.eventHandlers.set(eventName, [cfunc, []])
-
-      try {
-        this.vizzu._call(this.module._addEventListener)(cname, cfunc)
-      } finally {
-        this.module._free(cname)
-      }
     }
     this.eventHandlers.get(eventName)[1].push(handler)
   }
@@ -45,47 +47,68 @@ export default class Events {
     })
 
     if (handlers.length === 0) {
-      const cname = this.vizzu._toCString(eventName)
-      try {
-        this.vizzu._call(this.module._removeEventListener)(cname, cfunc)
-      } finally {
-        this.module._free(cname)
+      if (!this._isJSEvent(eventName)) {
+        const cname = this.vizzu._toCString(eventName)
+        try {
+          this.vizzu._callOnChart(this.module._removeEventListener)(cname, cfunc)
+        } finally {
+          this.module._free(cname)
+        }
+        this.module.removeFunction(cfunc)
       }
-      this.module.removeFunction(cfunc)
       this.eventHandlers.delete(eventName)
     }
   }
 
-  _invoke(eventName, param) {
+  _invoke(eventName, param, eventPtr = 0) {
+    const state = { canceled: false }
     try {
       if (this.eventHandlers.has(eventName)) {
-        const jsparam = this.vizzu._fromCString(param)
-
         for (const handler of [...this.eventHandlers.get(eventName)[1]]) {
-          const eventParam = JSON.parse(jsparam)
-          eventParam.preventDefault = () => {
-            this.vizzu._call(this.module._event_preventDefault)()
-          }
-          if (eventParam.data?.markerId) {
-            eventParam.data.getMarker = this._getMarkerProxy(eventParam.data.markerId)
-          }
-          if (eventParam.type.endsWith('-draw') || eventParam.type.startsWith('draw-')) {
-            eventParam.renderingContext = this.vizzu.render.dc()
-          }
-
+          const eventParam = this._isJSEvent(eventName)
+            ? this._makeJSEventParam(param, state)
+            : this._makeCEventParam(eventPtr, param, state)
           handler(eventParam)
         }
       }
     } catch (e) {
       console.log('exception in event handler: ' + e)
     }
+    return state.canceled
+  }
+
+  _isJSEvent(eventName) {
+    return eventName.startsWith('api-')
+  }
+
+  _makeJSEventParam(param, state) {
+    param.preventDefault = () => {
+      state.canceled = true
+    }
+    return param
+  }
+
+  _makeCEventParam(eventPtr, cString, state) {
+    const json = this.vizzu._fromCString(cString)
+    const param = JSON.parse(json)
+    param.preventDefault = () => {
+      this.vizzu._call(this.module._event_preventDefault)(eventPtr)
+      state.canceled = true
+    }
+    if (param.data?.markerId) {
+      param.data.getMarker = this._getMarkerProxy(param.data.markerId)
+    }
+    if (param.type.endsWith('-draw') || param.type.startsWith('draw-')) {
+      param.renderingContext = this.vizzu.render.dc()
+    }
+    return param
   }
 
   _getMarkerProxy(markerId) {
     let markerData = null
     return () => {
       if (!markerData) {
-        const cStr = this.vizzu._call(this.vizzu.module._chart_markerData)(markerId)
+        const cStr = this.vizzu._callOnChart(this.vizzu.module._chart_markerData)(markerId)
         markerData = JSON.parse(this.vizzu.module.UTF8ToString(cStr))
       }
       return markerData