+++ src/apps/weblib/js-api/render.js
@@ -1,5 +1,6 @@
 export default class Render {
-  init(update, canvas, log) {
+  init(ccanvas, update, canvas, log) {
+    this.ccanvas = ccanvas
     this.enabled = true
     this.polygonFirstPoint = false
     this.offscreenCanvas = document.createElement('CANVAS')
@@ -71,16 +72,26 @@ export default class Render {
     this.prevUpdateHash = hash
   }
 
-  updateFrame(force) {
+  updateFrame(force = false) {
     const start = performance.now()
     this.updateCanvasSize()
     if (this.mainCanvas.width > 0 && this.mainCanvas.height > 0) {
       const renderControl = !this.enabled ? 2 : force ? 1 : 0
-      this.update(this.cssWidth, this.cssHeight, renderControl)
+      this.update(this.ccanvas.id, this.cssWidth, this.cssHeight, renderControl)
     }
     const time = performance.now() - start
     if (this.log && time > 1) {
       console.log('Render.updateFrame: ' + time.toFixed(2) + 'ms')
+    }
+  }
+
+  clientToRenderCoor(clientPos) {
+    const rect = this.clientRect()
+    const scaleX = rect.width / (this.mainCanvas.width / this.scaleFactor)
+    const scaleY = rect.height / (this.mainCanvas.height / this.scaleFactor)
+    return {
+      x: (clientPos.x - rect.x) / scaleX,
+      y: (clientPos.y - rect.y) / scaleY
     }
   }
 }